WpfGfx: Stub out __if_exists.
__if_exists only exists in MSVC so I have to replace it somehow.

One would assume that __if_exists is used for a reason, and this
isn't going to be good enough. That's probably true. However, this
is a safe option, in that it won't disable any code that needs to
be included, and any code that should be disabled would presumably
fail to compile.

I don't see a way I can replace __if_exists with a working #define,
and there are still a lot of other errors in the headers. This
should at least delay the __if_exists errors until templates are
instantiated and I'm more able to tell if the fixes I make are
correct.