WpfGfx: Leave out some shader resources.
Having these in leads to attempts to compile shaders, which
crash because d3dcompiler is unimplemented. I don't really
understand why it works when the resources are absent, but since
this is how it worked before, it seems like the best compromise
until we have d3dcompiler working in Wine.